# docs/api/backoffice.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "helpaffe Backoffice API",
    "version": "0.1.0",
    "description": "The shared human and named-agent contract for support work. Browser-only access and credential administration operations are documented separately in docs/api.md."
  },
  "servers": [{ "url": "/api/backoffice" }],
  "security": [{ "browserSession": [] }, { "agentToken": [] }],
  "paths": {
...
    "/tickets": {
      "get": {
        "operationId": "listTickets",
        "summary": "List tickets in the caller's current project scope",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "$ref": "#/components/schemas/TicketStatus" } },
          { "name": "priority", "in": "query", "schema": { "$ref": "#/components/schemas/TicketPriority" } },
          { "name": "project_id", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "assignee_id", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "mine", "in": "query", "schema": { "type": "boolean", "default": false } },
          { "name": "search", "in": "query", "description": "PostgreSQL web-style full-text query across ticket metadata and conversation text", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Visible tickets", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TicketPage" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthenticated" }
        }
      }
    },
...
    "/tickets/{number}": {
      "parameters": [{ "$ref": "#/components/parameters/TicketNumber" }],
      "get": {
        "operationId": "getTicket",
        "summary": "Read one ticket with its complete support context",
        "responses": {
          "200": { "description": "Ticket, requester, instructions, and conversation", "headers": { "ETag": { "$ref": "#/components/headers/ETag" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TicketDetail" } } } },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "operationId": "updateTicket",
        "summary": "Change ticket status, priority, or assignee",
        "parameters": [{ "$ref": "#/components/parameters/IfMatch" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateTicket" } } } },
        "responses": {
          "200": { "description": "Updated ticket context", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TicketDetail" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "412": { "$ref": "#/components/responses/Stale" },
          "428": { "$ref": "#/components/responses/VersionRequired" }
        }
      }
    },
...
    "/tickets/{number}/replies": {
      "parameters": [{ "$ref": "#/components/parameters/TicketNumber" }],
      "post": {
        "operationId": "replyToTicket",
        "summary": "Atomically add a public reply and set its resulting status",
        "parameters": [{ "$ref": "#/components/parameters/IfMatch" }, { "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AddReply" } } } },
        "responses": {
          "200": { "description": "Updated or replayed ticket context", "headers": { "ETag": { "$ref": "#/components/headers/ETag" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TicketDetail" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "412": { "$ref": "#/components/responses/Stale" },
          "428": { "$ref": "#/components/responses/VersionRequired" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "browserSession": { "type": "apiKey", "in": "cookie", "name": "helpaffe_session" },
      "agentToken": { "type": "http", "scheme": "bearer", "bearerFormat": "hfa_ token" }
    },
    "parameters": {
      "TicketNumber": { "name": "number", "in": "path", "required": true, "schema": { "type": "string", "maxLength": 32 }, "example": "HLP-42" },
      "IfMatch": { "name": "If-Match", "in": "header", "required": true, "schema": { "type": "string", "pattern": "^\\\"[1-9][0-9]*\\\"$" }, "example": "\"8\"" },
      "IdempotencyKey": { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 1, "maxLength": 200 } },
...
    },
    "headers": {
      "ETag": { "description": "Quoted positive ticket version", "schema": { "type": "string" }, "example": "\"8\"" }
    },
    "responses": {
      "BadRequest": { "description": "Validation or cursor error", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
      "Unauthenticated": { "description": "Credential missing, unknown, or revoked", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
...
      "NotFound": { "description": "Resource absent, outside project scope, or no eligible next ticket", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
      "Conflict": { "description": "An idempotency key was reused for a different request", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
      "Stale": { "description": "The ticket version changed", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/StaleProblem" } } } },
      "VersionRequired": { "description": "If-Match was omitted", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } }
    },
    "schemas": {
      "TicketStatus": { "type": "string", "enum": ["open", "in_progress", "waiting_for_customer", "resolved"] },
      "TicketPriority": { "type": "string", "enum": ["normal", "urgent"] },
      "ProjectSummary": {
        "type": "object",
        "required": ["id", "key", "name"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "key": { "type": "string" },
          "name": { "type": "string" }
        },
        "additionalProperties": false
      },
      "UserSummary": {
        "type": ["object", "null"],
        "required": ["id", "name"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" }
        },
        "additionalProperties": false
      },
...
      "TicketSummary": {
        "type": "object",
        "required": ["id", "number", "project", "subject", "status", "priority", "version", "assignee", "created_at", "updated_at", "last_customer_reply_at", "waiting_since"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "number": { "type": "string" },
          "project": { "$ref": "#/components/schemas/ProjectSummary" },
          "subject": { "type": "string" },
          "status": { "$ref": "#/components/schemas/TicketStatus" },
          "priority": { "$ref": "#/components/schemas/TicketPriority" },
          "version": { "type": "integer", "minimum": 1 },
          "assignee": { "$ref": "#/components/schemas/UserSummary" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "last_customer_reply_at": { "type": "string", "format": "date-time" },
          "waiting_since": { "type": "string", "format": "date-time" }
        },
        "additionalProperties": false
      },
      "TicketPage": {
        "type": "object",
        "required": ["items", "next_cursor"],
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/TicketSummary" } },
          "next_cursor": { "type": ["string", "null"] }
        },
        "additionalProperties": false
      },
      "Requester": {
        "type": "object",
        "required": ["external_user_id", "name", "email"],
        "properties": {
          "external_user_id": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" }
        },
        "additionalProperties": false
      },
      "Actor": {
        "type": ["object", "null"],
        "required": ["user_id", "user_name", "agent_id", "agent_name"],
        "properties": {
          "user_id": { "type": "string", "format": "uuid" },
          "user_name": { "type": ["string", "null"] },
          "agent_id": { "type": ["string", "null"], "format": "uuid" },
          "agent_name": { "type": ["string", "null"] }
        },
        "additionalProperties": false
      },
      "ConversationEntry": {
        "type": "object",
        "required": ["id", "sequence", "kind", "body", "is_public", "created_at", "actor"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "sequence": { "type": "integer", "minimum": 1 },
          "kind": { "type": "string", "enum": ["customer_message", "public_reply", "internal_note", "system_event"] },
          "body": { "type": "string" },
          "is_public": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" },
          "actor": { "$ref": "#/components/schemas/Actor" }
        },
        "additionalProperties": false
      },
      "NotificationDelivery": {
        "type": "object",
        "required": ["id", "type", "target_kind", "recipient_email", "status", "attempt_count", "next_attempt_at", "submitted_at", "last_error", "created_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "type": { "$ref": "#/components/schemas/EmailTemplateType" },
          "target_kind": { "type": "string", "enum": ["customer", "assignee", "support_recipients"] },
          "recipient_email": { "type": ["string", "null"], "format": "email", "description": "Null until a support-recipient job is expanded" },
          "status": { "type": "string", "enum": ["pending", "submitted_to_smtp", "failed"] },
          "attempt_count": { "type": "integer", "minimum": 0 },
          "next_attempt_at": { "type": ["string", "null"], "format": "date-time" },
          "submitted_at": { "type": ["string", "null"], "format": "date-time" },
          "last_error": { "type": ["string", "null"] },
          "created_at": { "type": "string", "format": "date-time" }
        },
        "additionalProperties": false
      },
      "TicketDetail": {
        "type": "object",
        "required": ["summary", "requester", "context", "support_instructions", "notifications", "conversation"],
        "properties": {
          "summary": { "$ref": "#/components/schemas/TicketSummary" },
          "requester": { "$ref": "#/components/schemas/Requester" },
          "context": { "type": ["object", "null"], "description": "Immutable JSON snapshot supplied when the product created the ticket", "additionalProperties": true },
          "support_instructions": { "type": "string" },
          "notifications": { "type": "array", "items": { "$ref": "#/components/schemas/NotificationDelivery" } },
          "conversation": { "type": "array", "items": { "$ref": "#/components/schemas/ConversationEntry" } }
        },
        "additionalProperties": false
      },
      "UpdateTicket": {
        "type": "object",
        "properties": {
          "status": { "$ref": "#/components/schemas/TicketStatus" },
          "priority": { "$ref": "#/components/schemas/TicketPriority" },
          "assignee_id": { "type": ["string", "null"], "format": "uuid" },
          "clear_assignee": { "type": "boolean", "default": false }
        },
        "additionalProperties": false
      },
      "AddReply": {
        "type": "object",
        "required": ["message", "status"],
        "properties": {
          "message": { "type": "string", "minLength": 1 },
          "status": { "type": "string", "enum": ["in_progress", "waiting_for_customer", "resolved"] }
        },
        "additionalProperties": false
      },
...
      "EmailTemplateType": {
        "type": "string",
        "enum": ["new_ticket_customer", "new_ticket_support", "customer_reply_support", "public_reply_customer", "assignment_support"]
      },
...
      "Problem": {
        "type": "object",
        "required": ["type", "title", "status", "detail"],
        "properties": {
          "type": { "type": "string" },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "detail": { "type": "string" },
          "instance": { "type": "string" }
        }
      },
      "StaleProblem": {
        "allOf": [
          { "$ref": "#/components/schemas/Problem" },
          { "type": "object", "required": ["current_version"], "properties": { "current_version": { "type": "integer", "minimum": 1 } } }
        ]
      }
    }
  }
}